+++ src/components/profile/profile-more.tsx
@@ -186,7 +186,6 @@ export const ProfileMore = (props: Props) => {
           initialData={() => readProfileFile(id)}
           dataKey={id}
           language={id === "Merge" ? "yaml" : "javascript"}
-          schema={id === "Merge" ? "merge" : undefined}
           logInfo={id === "Script" ? entries : undefined}
           onSave={async (prev, curr) => {
             await saveProfileFile(id, curr ?? "");